Use JSON in locale files
