+++ package.json
@@ -11,7 +11,7 @@
     "staging": "vue-cli-service build --mode staging"
   },
   "dependencies": {
-    "@brazilian-utils/validators": "^0.1.9",
+    "@brazilian-utils/brazilian-utils": "^1.0.0-rc.4",
     "@fortawesome/fontawesome-free": "^5.11.2",
     "@simpli/meta-schema": "^1.0.2",
     "@simpli/resource-collection": "^1.0.7",
