+++ src/app/dialog/Dialog.ts
@@ -1,4 +1,4 @@
-import {$} from '@/simpli'
+import {$} from '@/facade'
 
 export class Dialog {
   title: string = ''
