+++ src/store/RootStore.ts
@@ -1,6 +1,6 @@
 import {ActionContext, ModuleTree, StoreOptions} from 'vuex'
 import {AccessorHandler} from '@simpli/vuex-typescript'
-import {PluginSetup} from '@/setup/PluginSetup'
+import {Setup} from '@/app/Setup'
 import {RootState} from '@/types/store'
 import {Lang} from '@/enums/Lang'
 import {Currency} from '@/enums/Currency'
@@ -32,7 +32,7 @@ export class RootStore implements StoreOptions<RootState> {
      * Change app language
      */
     setLanguage(context: RootContext, lang: Lang | null) {
-      if (lang) PluginSetup.changeLocale(lang)
+      if (lang) Setup.changeLocale(lang)
       context.commit('SET_LANGUAGE', lang)
     },
 
@@ -40,7 +40,7 @@ export class RootStore implements StoreOptions<RootState> {
      * Change app currency
      */
     setCurrency(context: RootContext, currency: Currency | null) {
-      if (currency) PluginSetup.changeCurrency(currency)
+      if (currency) Setup.changeCurrency(currency)
       context.commit('SET_CURRENCY', currency)
     },
   }
