+++ src/types/v-money.d.ts
@@ -1,0 +1,5 @@
+declare module 'v-money' {
+  import {PluginFunction} from 'vue/types/plugin'
+  const Plugin: PluginFunction<{}>
+  export default Plugin
+}
