+++ src/types/vue-moment.d.ts
@@ -1,5 +1,5 @@
 declare module 'vue-moment' {
   import {PluginFunction} from 'vue/types/plugin'
-  const VueMoment: PluginFunction<{}>
-  export default VueMoment
+  const Plugin: PluginFunction<{}>
+  export default Plugin
 }
