+++ src/types/vue-the-mask.d.ts
@@ -1,0 +1,5 @@
+declare module 'vue-the-mask' {
+  import {PluginFunction} from 'vue/types/plugin'
+  const Plugin: PluginFunction<{}>
+  export default Plugin
+}
