+++ src/types/vue-transition-expand.d.ts
@@ -1,0 +1,5 @@
+declare module 'vue-transition-expand' {
+  import {PluginFunction} from 'vue/types/plugin'
+  const Plugin: PluginFunction<{}>
+  export default Plugin
+}
